+++ app/Models/Order.php
@@ -2,6 +2,10 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use App\Models\PackagingQueue;
+use App\Models\ShipmentQueue;
+use App\Models\Employee;
+use App\Models\OrderedBook;
 
 class Order extends Model
 {
@@ -31,6 +35,32 @@ class Order extends Model
         return $this->belongsTo(Employee::class, 'handled_by', 'id');
     }
 
+    public function designQueue()
+    {
+        return $this->hasOne(DesignQueue::class, 'ordered_book_id', 'ordered_book_id');
+    }
+
+    public function printingQueue()
+    {
+        return $this->hasOne(PrintingQueue::class, 'ordered_book_id', 'ordered_book_id' );
+    }
+
+    public function coverPrintingQueue()
+    {
+        return $this->hasOne(CoverPrintingQueue::class, 'ordered_book_id', 'ordered_book_id');
+    }
+
+    public function bindingQueue()
+    {
+        return $this->hasOne(BindingQueue::class, 'ordered_book_id', 'ordered_book_id');
+    }
+
+    public function qcQueue()
+    {
+        return $this->hasOne(QcQueue::class, 'ordered_book_id', 'ordered_book_id');
+    }
+
+
     // Relationships for packaging and shipment queue
     public function packagingQueue()
     {
@@ -41,4 +71,6 @@ class Order extends Model
     {
         return $this->hasOne(ShipmentQueue::class, 'order_id', 'order_id');
     }
+
+
 }